refactor: extract transaction-related code to new txn.go file
